privatize IConnection and HttpConnection
