+++ clients/ts/signalr/spec/HttpClient.spec.ts
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-import { HttpRequest } from "../src/index";
+import { HttpRequest } from "../src/HttpClient";
 import { TestHttpClient } from "./TestHttpClient";
 import { asyncit as it } from "./Utils";
