+++ clients/ts/signalr/spec/HubConnection.spec.ts
@@ -1,11 +1,12 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-import { HubConnection, JsonHubProtocol } from "../src/HubConnection";
+import { HubConnection } from "../src/HubConnection";
 import { IConnection } from "../src/IConnection";
 import { HubMessage, IHubProtocol, MessageType } from "../src/IHubProtocol";
 import { ILogger, LogLevel } from "../src/ILogger";
 import { HttpTransportType, ITransport, TransferFormat } from "../src/ITransport";
+import { JsonHubProtocol } from "../src/JsonHubProtocol";
 import { NullLogger } from "../src/Loggers";
 import { IStreamSubscriber } from "../src/Stream";
 import { TextMessageFormat } from "../src/TextMessageFormat";
@@ -13,7 +14,7 @@ import { TextMessageFormat } from "../src/TextMessageFormat";
 import { asyncit as it, captureException, delay, PromiseSource } from "./Utils";
 
 function createHubConnection(connection: IConnection, logger?: ILogger, protocol?: IHubProtocol) {
-    return new HubConnection(connection, logger || NullLogger.instance, protocol || new JsonHubProtocol());
+    return HubConnection.create(connection, logger || NullLogger.instance, protocol || new JsonHubProtocol());
 }
 
 describe("HubConnection", () => {
