+++ clients/ts/signalr/src/HubConnection.ts
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 import { HandshakeProtocol, HandshakeRequestMessage, HandshakeResponseMessage } from "./HandshakeProtocol";
-import { HttpConnection, IHttpConnectionOptions } from "./HttpConnection";
+import { HttpConnection } from "./HttpConnection";
 import { IConnection } from "./IConnection";
 import { CancelInvocationMessage, CompletionMessage, HubMessage, IHubProtocol, InvocationMessage, MessageType, StreamInvocationMessage, StreamItemMessage } from "./IHubProtocol";
 import { ILogger, LogLevel } from "./ILogger";
@@ -11,8 +11,6 @@ import { NullLogger } from "./Loggers";
 import { IStreamResult } from "./Stream";
 import { TextMessageFormat } from "./TextMessageFormat";
 import { Arg, createLogger, Subject } from "./Utils";
-
-export { JsonHubProtocol };
 
 const DEFAULT_TIMEOUT_IN_MS: number = 30 * 1000;
 
@@ -30,7 +28,14 @@ export class HubConnection {
 
     public serverTimeoutInMilliseconds: number;
 
-    constructor(connection: IConnection, logger: ILogger, protocol: IHubProtocol) {
+    /** @internal */
+    // Using a public static factory method means we can have a private constructor and an _internal_
+    // create method that can be used by HubConnectionBuilder.
+    public static create(connection: IConnection, logger: ILogger, protocol: IHubProtocol): HubConnection {
+        return new HubConnection(connection, logger, protocol);
+    }
+
+    private constructor(connection: IConnection, logger: ILogger, protocol: IHubProtocol) {
         Arg.isRequired(connection, "connection");
         Arg.isRequired(logger, "logger");
         Arg.isRequired(protocol, "protocol");
