+++ clients/ts/signalr/src/IHttpConnectionOptions.ts
@@ -1,0 +1,15 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+import { HttpClient } from "./HttpClient";
+import { ILogger, LogLevel } from "./ILogger";
+import { HttpTransportType, ITransport } from "./ITransport";
+
+export interface IHttpConnectionOptions {
+    httpClient?: HttpClient;
+    transport?: HttpTransportType | ITransport;
+    logger?: ILogger | LogLevel;
+    accessTokenFactory?: () => string | Promise<string>;
+    logMessageContent?: boolean;
+    skipNegotiation?: boolean;
+}
